pyproject.toml
[tool.poetry]
name = "vision-ai-lab"
version = "0.1.0"
description = ""
authors = ["sangju.lee <sjleee.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
opencv-python = "^4.10.0.84"
coremltools = "^7.2"
tensorflow = "^2.17.0"
pillow = "^10.4.0"
ultralytics = "^8.2.75"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
